requirements.txt
# Core dependencies
fastapi==0.68.1
uvicorn==0.15.0
pydantic==1.8.2
starlette==0.14.2

# HTTP and API
httpx==0.19.0
requests==2.26.0
python-multipart==0.0.5
python-jose[cryptography]==3.3.0

# Environment and configuration
python-dotenv==0.19.0

# Payment integration
stripe==2.60.0

# Monitoring and metrics
prometheus-client==0.11.0
sentry-sdk[fastapi]==1.3.1

# Caching
redis==3.5.3

# Testing
pytest==6.2.5
pytest-asyncio==0.15.1
locust==2.2.1

# Additional dependencies
h11==0.12.0
httpcore==0.13.7
asgiref==3.8.1
charset-normalizer==2.0.12
cffi==1.17.1 